nyx-modules-qemux86: Fix '-EINVAL' error for vboxguest ioctl
:Release Notes:
Fix '-EINVAL' error for vboxguest ioctl

:Detailed Notes:
The implementation code of vboxguest included in the kernel is slightly
different from the implementation of vboxguest included in VirtualBox.
Therefore, the processing for the vboxguest ioctl is also slightly
different.

Requesting the VMMDevReq_SetMouseStatus, VMMDevReq_SetPointerShape
request to the kernel's vboxguest driver causes an -EINVAL error.
Kernel code has the following annotations:

/ * IOCTL to perform a VMM Device request less than 1KB in size. * /
VBG_IOCTL_VMMDEV_REQUEST(s) _IOC(_IOC_READ|_IOC_WRITE, 'V', 2, s)

/ * IOCTL to perform a VMM Device request larger then 1KB. * /
VBG_IOCTL_VMMDEV_REQUEST_BIG _IOC(_IOC_READ|_IOC_WRITE, 'V', 3, 0)

According to the annotation, the request to be called in nyx is smaller
than 1KB, so it is correct to transmit it as 'V', 2. (Not 'V', 3)
An EINVAL error does not occur in this modification.

:Testing Performed:
1. You can test the init_vbox_touchpanel () function by creating a
separate main file.
2. Build kernel to include vboxguest driver
3. Run webOS on VirtualBox
4. If the mouse cursor is shown on the webOS, it's OK.

:QA Notes:
N/A

:Issues Addressed:
[PLAT-77791] Support Virtualbox

Change-Id: Ia229bb0215099519e8c933a441c53883751c84ed